Use id as symbol for binance dapi
Binance spot and fapi have zero or one market for each base/quote pair.
Binance dapi can have more than one market. As such, we should use the
market identifier for each symbol. This is consistent with previous
examples in ccxt, and the discussions referenced from ccxt/ccxt#7122.

It would be possible to use the unified symbols for the perpetual
contracts on Binance dapi; this commit does not do so.